Add --download-art to get art for shortcuts
Downloads art from Steam into Steam's expected folder structure. Just
putting the art there is enough to make it show up in the client.
Sets art for Steam desktop and Big Picture (instead of the icons).

Use Steam's GetAppList to get all games so we can lookup appids. Cache
it and expire after a week. Games are usually added to Steam before
release, so this seems like a good compromise of hitting the server and
getting relevant data.

The biggest challenge is figuring out the shortcut -> image filename
mapping so that Steam recognizes the downloaded image as art. Follows
the steamgrid's method to determine the id for the shortcut to write the
correct filenames:
  https://github.com/boppreh/steamgrid/blob/c796e612c67925413317f4012bdc771326f173c8/games.go#L100-L137
For some reason, they refer it as the LegacyID, but it seems to be the
future. This id seems to be unrelated to the desktop shortcut id (which
is also used in screenshots.vdf) and unrelated to images in Big Picture
mode (which uses icon in shortcuts.vdf).

Big Picture uses the same ids, but 64 bit with some constant stuck in
the lower bits:
    https://github.com/scottrice/Ice/blob/7130b54c8d2fa7d0e2c0994ca1f2aa3fb2a27ba9/ice/steam_grid.py#L49-L64

Test
* Download art for Spiritfarer.