Add --download-art-all-shortcuts
Add a flag for downloading art for shortcuts we aren't specifying. It
will download for shortcuts we don't create and shortcuts we created in
the past but aren't selecting right now.

Specifying --download-art-all-shortcuts downloads all art regardless of
the current game selection.

Change how we pass data to add_games_to_shortcut_file so we have access
to SteamDatabase and SteamAccount.