Try harder to detect xbox games: remove id filter
Fix GenesisNoir, Supraland not detected as games.

Don't filter by manifest id. Seems that lots of apps put whatever they
want in there -- often the app name.

These two games put something like their app name as the id:
    Supraland: AppSupralandShipping
    GenesisNoir: GenesisNoir

Instead, include everything and rely on our other filtering to determine
if it's a game. We still pass along the id because if it's 'Game', then
it's probably a game.