Add --dump-shortcut-vdf argument
Having a simple way to view the vdf file makes debugging shortcut issues
much easier. Add an argument so steamsync can do this for us.

I had a script similar to this that I used while debugging art
downloads. Putting it inside steamsync ensures it's printing the
shortcuts file we're working with (for the right user).