Ignore utf errors in xbox manifests
Fix failure to process xbox games because minidom can't parse Copilot's
manifest:
    UnicodeDecodeError: 'utf-8' codec can't decode byte 0xff in position 0: invalid start byte

Adding a print before the minidom.parse shows that the failure was from
Copilot:
    <_io.TextIOWrapper name='C:\\Program Files\\WindowsApps\\Microsoft.Copilot_1.1.9.0_neutral__8wekyb3d8bbwe\\AppxManifest.xml' mode='r' encoding='utf-8'>

Ignore errors when loading the manifest so it's more likely we can read
it. If it ends up being unparsable, then the manifest failed to tell us
it's a game and that's fine.

Alternatively, we could catch this specific error but I don't think the
failure really matters.

Test
* xbox import now completes successfully.
* Changed parse line to force parse failure:
    doc = minidom.parseString("malformedxml"+f.read())
  See output and games continue to import correctly:
    Failed to parse manifest and assuming not a game: 'C:/Program Files/WindowsApps/Microsoft.Xbox.TCUI_1.24.10001.0_x64__8wekyb3d8bbwe/AppxManifest.xml'